delete DoctorSpecialization.java, change the logic of adding doctor(relate doctor with Specialization), change nullable values in user, doctor functions done ,add devtools to pom

+++ src/main/java/com/grad/akemha/service/userServive/AdminService.java
@@ -1,50 +1,66 @@
 package com.grad.akemha.service.userServive;
 
 import com.grad.akemha.dto.AddDoctorDto;
-import com.grad.akemha.entity.DoctorSpecialization;
+import com.grad.akemha.entity.Specialization;
 import com.grad.akemha.entity.User;
-import com.grad.akemha.repository.DoctorSpecializationRepository;
+import com.grad.akemha.repository.SpecializationRepository;
 import com.grad.akemha.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
+@RequiredArgsConstructor
 public class AdminService {
     @Autowired
     UserRepository userRepository;
 
     @Autowired
-    DoctorSpecializationRepository doctorSpecializationRepository;
+    SpecializationRepository specializationRepository;
+
+    private final PasswordEncoder passwordEncoder;
+
 
     public List<User> getDoctors() {
         return userRepository.findAll();
     }
 
-    public void addDoctor(AddDoctorDto request) {
+    public User addDoctor(AddDoctorDto request) {
         User user = new User();
         user.setName(request.getName());
         user.setEmail(request.getEmail());
-        user.setEmail(request.getEmail());
+        user.setPassword(passwordEncoder.encode(request.getPassword()));
         user.setGender(request.getGender());
         //check if the specialization not exist  create it and assign it to user
-        DoctorSpecialization doctorSpecialization = doctorSpecializationRepository.findBySpecializationType(request.getSpecialization());
-        if (doctorSpecialization==null) {
-            doctorSpecialization.setSpecializationType(request.getSpecialization());
-            doctorSpecializationRepository.save(doctorSpecialization);
+        Specialization specialization = specializationRepository.findBySpecializationType(request.getSpecialization());
+        System.out.println(specialization);
+        if (specialization == null) {
+            specialization=new Specialization();
+            specialization.setSpecializationType(request.getSpecialization());
+            specializationRepository.save(specialization);
         }
-        user.setDoctorSpecialization(doctorSpecialization);
-        userRepository.save(user);
+        user.setSpecialization(specialization);
+        return userRepository.save(user);
     }
 
-    public void deleteDoctor(Long userId) {
-        userRepository.deleteById(userId);
+    public boolean deleteDoctor(Long userId) {
+        Optional<User> optionalUser = userRepository.findById(userId);
+        if (optionalUser.isPresent()) {
+            userRepository.deleteById(userId);
+            return true;
+        }
+        return false;
     }
 
-    public List<DoctorSpecialization> getDoctorSpecialization() {
-        return doctorSpecializationRepository.findAll();
+    public List<Specialization> getDoctorSpecialization() {
+        return specializationRepository.findAll();
     }
 
 
 }
+
+
